% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_summary_table.R
\name{build_summary_table}
\alias{build_summary_table}
\title{Generates an empty summary table (as a tibble ready for storing log records (1 for each batch chunk processed)).}
\usage{
build_summary_table()
}
\value{
A tibble (data frame) ready for storing data fetching log records (with the predefined structure ready for inserting in the student_miguel.pipeline_logs DB table).
}
\description{
Generates an empty summary table (as a tibble ready for storing log records (1 for each batch chunk processed)).
}
\examples{
\dontrun{
tbl <- build_summary_table()
}
}
