% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_symbols.R
\name{fetch_symbols}
\alias{fetch_symbols}
\title{Obtains a list of index_ts and symbol pairs}
\usage{
fetch_symbols(conn)
}
\arguments{
\item{conn}{A (DBI/Postgres) connection to the database from which reference data is read and result data is written to.}
}
\value{
A data frame containing the pairs (index_ts, symbol) found in the reference data table; or NULL, in case of a non-valid DB connection.
}
\description{
Obtains a list of index_ts and symbol pairs
}
\examples{
\dontrun{
conn <- connect_db()
symbols <- fetch_symbols(conn)
}
}
