% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Re-structures financial data received from Yahoo!Finance API call into a structure directly compatible with DB table student_miguel.data_sp500 (for straightforward appending of records).}
\usage{
format_data(fin_data, symbols)
}
\arguments{
\item{fin_data}{The financial data (a data frame) obtained from Yahoo!Finance.}

\item{symbols}{A set of company symbols and associated index_ts, typically covering all the same symbols present in the financial data.}
}
\value{
A data frame containing the same data (as fin_data) but with the required pre-defined structure / format.
}
\description{
Re-structures financial data received from Yahoo!Finance API call into a structure directly compatible with DB table student_miguel.data_sp500 (for straightforward appending of records).
}
\examples{
\dontrun{
df <- data.frame(symbol = c("AAPL", "META"), index_ts = c("apple_inc_aapl", "meta_platforms_meta"))
data_restructured <- format_data(fin_data, df)
}
}
