% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_new_data.R
\name{insert_new_data}
\alias{insert_new_data}
\title{Appends a group of financial records (corresponding to a chunk) to the student_miguel.data_sp500 DB table; all records are tested for duplicates (with regard to data already in the DB table).}
\usage{
insert_new_data(conn, next_id = 0, formatted_data)
}
\arguments{
\item{conn}{A DB connection.}

\item{next_id}{The next valid / available record identifier (an integer) to use as the starting unique identifier for the (first record of the) group of records to be inserted; if not specified, it is zero by default, which will trigger its automatic determination based on the last (highest value) currently in the DB table.}

\item{formatted_data}{The group of financial records to be inserted (already formatted for direct appending to the DB table -- as per format_data()).}
}
\value{
The quantity of records actually inserted in the DB, or, otherwise, the value -1 (in case a DB connection is not valid when needed).
}
\description{
Appends a group of financial records (corresponding to a chunk) to the student_miguel.data_sp500 DB table; all records are tested for duplicates (with regard to data already in the DB table).
}
\examples{
\dontrun{
result <- insert_new_data(conn, formatted_data = formatted_data)
result <- insert_new_data(conn, 0, formatted_data)
result <- insert_new_data(conn, 999999999, formatted_data)
}
}
