% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_summary.R
\name{log_summary}
\alias{log_summary}
\title{Generates and stores a (log) record with details about the processing of batch chunk in the Summary Table (a tibble).}
\usage{
log_summary(batch_log, user_login, chunk_id, symbols, rows_inserted)
}
\arguments{
\item{batch_log}{The existing Summary Table (as a tibble -- should be created by build_summary_table() or updated by this function).}

\item{user_login}{The username (a string) of the user connecting to the DB.}

\item{chunk_id}{The chunk count identifier (for the batch): an integer starting at 1.}

\item{symbols}{The list of company symbols and corresponding index_ts. (index_ts data is missing in the examples and in the tests!)}

\item{rows_inserted}{The quantity of financial records inserted in the DB.}
}
\value{
The updated version of the Summary Table, with the log record added; a record is added to the log regardless of if financial records are added or not to the DB (in the student_miguel.data_sp500 table).
}
\description{
Generates and stores a (log) record with details about the processing of batch chunk in the Summary Table (a tibble).
}
\examples{
\dontrun{
log <- log_summary(log, "miguel", 1, data.frame(symbol = c("AAPL", "META")), 4)
log <- log_summary(log, "miguel", 5, data.frame(symbol = c("AAPL", "META")), 0)
}
}
