% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_summary_table.R
\name{push_summary_table}
\alias{push_summary_table}
\title{Appends a summary table to the student_miguel.pipeline_logs DB table.}
\usage{
push_summary_table(conn, batch_log)
}
\arguments{
\item{conn}{A DB connection.}

\item{batch_log}{The summary table (as a tibble) containing the records to be appended (and using a structure matching the DB table); this object should be created using build_summary_table() and populated using log_summary().}
}
\value{
TRUE, if the append operation was successful; FALSE, otherwise.
}
\description{
Appends a summary table to the student_miguel.pipeline_logs DB table.
}
\examples{
\dontrun{
result <- push_summary_table(conn, batch_log)
}
}
