% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_batch.R
\name{split_batch}
\alias{split_batch}
\title{Divides the list of symbols (the batch) into separate sections (chunks), so that the Yahoo!Finance API does not get overloaded when the calls are made.}
\usage{
split_batch(batch, chunk_size)
}
\arguments{
\item{batch}{A data frame containing a list of symbols (with at least 1 column named "symbol").}

\item{chunk_size}{The size (an integer greater than zero) of each section (chunk).}
}
\value{
A list of data frames containing each 1 column (named "symbol") of size equal to chunk_size; in case chunk_size is invalid, then a default value of 20 is assumed.
}
\description{
Divides the list of symbols (the batch) into separate sections (chunks), so that the Yahoo!Finance API does not get overloaded when the calls are made.
}
\examples{
\dontrun{
batch_chunks <- split_batch(batch, 50)
}
}
