% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_query_data.R
\name{yahoo_query_data}
\alias{yahoo_query_data}
\title{Obtains financial data for the set of company symbols provided for the date period specified (including date_from, but excluding date_to).}
\usage{
yahoo_query_data(chunk, date_from, date_to)
}
\arguments{
\item{chunk}{Set of company symbols (data frame of 1 column named "symbol").}

\item{date_from}{Date (string in format "YYYY-MM-DD") since when financial data should be obtained.}

\item{date_to}{Date (string in format "YYYY-MM-DD") until when financial data should be obtained (data is only obtained until the day before the date provided).}
}
\value{
A data frame containing the financial data as obtained from the call to Yahoo!Finance's API.
}
\description{
Obtains financial data for the set of company symbols provided for the date period specified (including date_from, but excluding date_to).
}
\examples{
\dontrun{
fin_data <- yahoo_query_data(data.frame(symbol = c("AAPL", "META")), "2024-04-01", "2024-04-03")
fin_data <- yahoo_query_data(data.frame(symbol = c("AAPL", "META")), "2024-04-04", "2024-04-05")
}
}
