% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_pipeline.R
\name{start_pipeline}
\alias{start_pipeline}
\title{Runs a data-fetching iteration of the data pipeline}
\usage{
start_pipeline(start_date = Sys.Date() - 1, end_date = Sys.Date())
}
\arguments{
\item{start_date}{The date (as a string in the format YYYY-MM-DD) of the first day for which data is to be fetched. By default, it is the date of yesterday.}

\item{end_date}{The date (as a string in the format YYYY-MM-DD) of the day after the last day for which data is to be fetched. By default, it is the date of today.}
}
\value{
Nothing.
}
\description{
Runs a data-fetching iteration of the data pipeline
}
\examples{
\dontrun{
start_pipeline()
start_pipeline("2024-04-01", "2024-04-03")
start_pipeline("2024-04-01", "2024-04-30")
}
}
